refactor(app:ccf-organ-info): :recycle: refactor dir structure of deps in the component
